+++ source/include/Ostap/MoreMath.h
@@ -2557,39 +2557,55 @@ namespace Ostap
     // ========================================================================
     /// the sigmoid type 
     enum class SigmoidType
-      {
-	//
-	Logistic              , // based on logistic function 
-	Hyperbolic            , // based on tanh 
-	Trigonometric         , // based on atan 
-	Error                 , // Based on error function 
-	Gudermannian          , // Based on Gudermannian function
-	Algebraic             , // 0.5 * ( 1 + 2*x / hypot ( 1 , 2*x ) )
-	SmoothTransition      , // Based on "smooth transition" function
-	//
-	Polynomial_n0         , // Based on "smooth step" with n=0
-	Polynomial_n1         , // Based on "smooth step" with n=1
-	Polynomial_n2         , // Based on "smooth step" with n=2
-	Polynomial_n3         , // Based on "smooth step" with n=3
-	Polynomial_n4         , // Based on "smooth step" with n=4
-	Polynomial_n5         , // Based on "smooth step" with n=5
-	Polynomial_n6         , // Based on "smooth step" with n=6
-	//
-	First = Logistic      , 
-	Last  = Polynomial_n6 
-      } ;      	
+    {
+      //
+      Logistic              , // based on logistic function 
+      Hyperbolic            , // based on tanh 
+      Trigonometric         , // based on atan 
+      Error                 , // Based on error function 
+      Gudermannian          , // Based on Gudermannian function
+      Algebraic             , // 0.5 * ( 1 + 2*x / hypot ( 1 , 2*x ) )
+      SmoothTransition      , // Based on "smooth transition" function
+      //
+      Polynomial_n0         , // Based on "smooth step" with n=0
+      Polynomial_n1         , // Based on "smooth step" with n=1
+      Polynomial_n2         , // Based on "smooth step" with n=2
+      Polynomial_n3         , // Based on "smooth step" with n=3
+      Polynomial_n4         , // Based on "smooth step" with n=4
+      Polynomial_n5         , // Based on "smooth step" with n=5
+      Polynomial_n6         , // Based on "smooth step" with n=6
+      //
+      First = Logistic      , 
+      Last  = Polynomial_n6 
+    } ;      	
     // ========================================================================
-    /** Sigmoid function
+    /** sigmoid type
+     *  @param  name the case-insensitive name of of sigmoid function
+     *  @return ID   of sigmoid function
+     */
+    SigmoidType sigmoid_type
+    ( const std::string& name = "Hyperbolic" ) ;
+    // ========================================================================
+    /*  sigmoid type
+     *  @param  sigmoid ID 
+     *  @return the name of of sigmoid function
+     */
+    // ========================================================================
+    std::string sigmoid_name 
+    ( const SigmoidType stype ) ;
+    // ======================================================================= 
+    /** Sigmoid functions
      *  All sigmoid fuctions \f$ \sigma(z) \f$ are normalized & scaled such
      *  - \f$ \sigma(-\infty) =0\f$ 
      *  - \f$ \sigma(+\infty) =1\f$ 
-     *  - \f$ \sigma^\prime(0)=1\f$ 
+     *  - \f$ \sigma^\prime(0)=1\f$
+     * @see Ostap::Math::SigmoidType
      */
     double sigmoid
     ( const double      x                           ,
       const SigmoidType t = SigmoidType::Hyperbolic ) ;
     // ========================================================================
-    
+
     // ========================================================================
     // Moebius ransformation
     // ========================================================================